Remove dependence on dummy links in the database setup
* Also, slim down the Link tests now that `Link` objects have an `id`.

+++ Rakefile
@@ -4,14 +4,7 @@ task :test_database_setup do
   p "Cleaning database..."
 
   connection = PG.connect(dbname: 'bookmark_manager_test')
-
-  # Clear the database
-  connection.exec("TRUNCATE comments, links;")
-
-  # Add the test data
-  connection.exec("INSERT INTO links VALUES(1, 'http://www.makersacademy.com', 'Makers Academy');")
-  connection.exec("INSERT INTO links VALUES(2, 'http://www.google.com', 'Google');")
-  connection.exec("INSERT INTO links VALUES(3, 'http://www.facebook.com', 'Facebook');")
+  connection.exec("TRUNCATE comments, links, users;")
 end
 
 task :setup do
@@ -20,17 +13,23 @@ task :setup do
   ['bookmark_manager', 'bookmark_manager_test'].each do |database|
     connection = PG.connect
     connection.exec("CREATE DATABASE #{ database };")
+
     connection = PG.connect(dbname: database)
     connection.exec("CREATE TABLE links(id SERIAL PRIMARY KEY, url VARCHAR(60), title VARCHAR(60));")
     connection.exec("CREATE TABLE comments(id SERIAL PRIMARY KEY, link_id INTEGER REFERENCES links (id), text VARCHAR(240));")
+    connection.exec("CREATE TABLE users(id SERIAL PRIMARY KEY, email VARCHAR(60), password VARCHAR(140));")
   end
 end
 
 task :teardown do
-  p "Tearing down databases..."
+  p "Destroying databases...type 'y' to confirm that you want to destroy the Bookmark Manager databases. This will remove all data in those databases!"
+
+  confirm = STDIN.gets.chomp
+
+  return unless confirm == 'y'
 
   ['bookmark_manager', 'bookmark_manager_test'].each do |database|
     connection = PG.connect
-    connection.exec("DROP DATABASE #{ database };")
+    connection.exec("DROP DATABASE #{ database }")
   end
 end
